use get_spark_cluster to start spark cluster
